refactor class names, create projection_selector
