+++ pytripgui/app_logic/viewcanvas.py
@@ -1,11 +1,11 @@
-from pytripgui.viewcanvas_vc.viewcanvas_view import ViewCanvasView
-from pytripgui.viewcanvas_vc.viewcanvas_cont import ViewCanvasCont
+from pytripgui.canvas_vc.canvas_view import CanvasView
+from pytripgui.canvas_vc.canvas_controller import CanvasController
 
 
 class ViewCanvases:
     def __init__(self):
-        self.viewcanvas_view = ViewCanvasView()
-        self.plot_cont = ViewCanvasCont(None, self.viewcanvas_view)
+        self.viewcanvas_view = CanvasView()
+        self.plot_cont = CanvasController(None, self.viewcanvas_view)
 
     def widget(self):
         return self.viewcanvas_view.widget()
