+++ pytripgui/canvas_vc/objects/vc_text.py
