+++ pytripgui/canvas_vc/objects/vdx.py
